Remove Interactive abstraction
